add todo save to localstorage

+++ src/index.js
@@ -2,6 +2,10 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap';
 
+import { fetchLibrary, saveLibrary } from './fetchLibrary';
+
+export const todoList = [];
+
 export class Project {
   constructor(name) {
     this.name = name;
@@ -18,9 +22,11 @@ export class Todo {
   }
 }
 
-const addTodo = (todo) => {
-  const todoCtn = document.querySelector('#todo-list');
-  const cardHtml = `
+export const render = () => {
+  const todoCtn = document.querySelector('#todo-list'); // just appdend it
+  todoCtn.innerHTML = '';
+  todoList.map((todo) => {
+    const cardHtml = `
     <div class="col mb-4">
       <div class="card border-primary mb-3" style="max-width: 18rem;">
         <div class="card-header">Header</div>
@@ -32,12 +38,18 @@ const addTodo = (todo) => {
         <div class="card-footer">
           <small class="text-muted">${todo.dueDate}</small>
           <i class="fa fa-edit ml-auto" id="edit"></i>
-          <i class="fa fa-trash ml-auto" id="delete"></i>
+          <a href=""><i class="fa fa-trash ml-auto" id="delete"></i></a>
         </div>
       </div>
     </div>
   `;
-  todoCtn.innerHTML += cardHtml;
+    todoCtn.innerHTML += cardHtml;
+    return null;
+  });
+};
+
+export const addTodo = (todo) => {
+  todoList.push(todo);
 };
 
 const clearFields = () => {
@@ -48,13 +60,11 @@ const clearFields = () => {
 };
 
 const deleteTodo = (target) => {
-  if (target.className === 'delete') {
-    target.parentElement.parentElement.remove();
-  }
+  target.parentElement.parentElement.parentElement.parentElement.remove();
 };
 
 document.querySelector('#form').addEventListener('submit', (e) => {
-  e.preventDefault();
+  // e.preventDefault();
   const title = document.querySelector('#todo-title').value;
   const desc = document.querySelector('#todo-desc').value;
   const date = document.querySelector('#todo-date').value;
@@ -63,11 +73,25 @@ document.querySelector('#form').addEventListener('submit', (e) => {
   const todo = new Todo(title, desc, date, priority);
   addTodo(todo);
   clearFields();
+  render();
+  $('#exampleModal').modal('toggle');
+  // document.getElementById('#exampleModal').modal('hide');
 });
 
-document.getElementById('delete').addEventListener('cick', (e) => {
-  deleteTodo(e.target);
+document.getElementById('todo-list').addEventListener('click', (e) => {
   e.preventDefault();
+  if (e.target.id === 'delete') {
+    deleteTodo(e.target);
+    render();
+  }
 });
 
 // document.querySelector('#edit').addEventListener('cick', () => {});
+
+const todo = new Todo('test1', 'desc', '2021-08-08', 'low');
+addTodo(todo);
+render();
+// window.addEventListener('load', renderInit);
+
+window.addEventListener('load', fetchLibrary);
+window.addEventListener('unload', saveLibrary);
